+++ pydantic2protobuf/cli_app.py
@@ -2,7 +2,7 @@ import logging
 
 import click
 
-from pydantic2protobuf.services.gen_protobuf import gen_protobuf_from_routes
+from pydantic2protobuf.services.fastapi_to_proto import routes_to_proto
 
 
 @click.command()
@@ -10,7 +10,7 @@ def gen_proto_for_services():
     # from webserver.api.commands_with_grpc import router as commands_with_grpc_router
     # routes = commands_with_grpc_router.routes
     routes = []
-    print(gen_protobuf_from_routes(routes))
+    print(routes_to_proto(routes))
 
 
 @click.group(context_settings={"help_option_names": ["-h", "--help"]})
